changes to project table

+++ database/migrations/2021_04_05_000000_create_projects_table.php
@@ -18,7 +18,7 @@ class CreateProjectsTable extends Migration
             $table->string('Customer_Address');
             $table->date('Date_Requested') ;
             $table->string('Completed');
-            $table->integer('total_cost')->nullable();
+            $table->decimal('total_cost')->nullable();
             $table->date('date_completed')->nullable();
             $table->string('invoice_paid')->nullable();
             $table->timestamps();
